requirements.txt
alabaster==1.0.0
attrs==25.4.0
babel==2.17.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
docutils==0.21.2
flasgger==0.9.7.1
Flask==3.1.2
idna==3.11
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mistune==3.1.4
packaging==25.0
pillow==12.0.0
Pygments==2.19.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
roman-numerals-py==3.1.0
rpds-py==0.28.0
six==1.17.0
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
swagger-ui-py==25.7.1
urllib3==2.5.0
Werkzeug==3.1.3
